requirements.txt
langchain_openai
langchain_core
python-dotenv
streamlit
langchain_community
langserve
fastapi
uvicorn
sse_starlette
pyPDF2
chromadb
google-generativeai
bs4
faiss-cpu
langchain_google_genai
pypdf
arxiv
wikipedia
groq
langchain-groq
cassio
langchainhub